# src/main/kotlin/tech/arhr/quingo/dto/QuizEnums.kt
package tech.arhr.quingo.dto

enum class Visibility {
    PUBLIC,
    PRIVATE,
}

enum class QuizStatus {
    UNPUBLISHED,
    PUBLISHED,
}

enum class CardType {
    SINGLE_CHOICE,
    MULTIPLE_CHOICE,
    TEXT_INPUT,
}